Add RF feature importance plotting
