+++ workflows/offline_ml_diags/offline_ml_diags/compute_diags.py
@@ -23,13 +23,14 @@ from vcm import safe, interpolate_to_pressure_levels
 import vcm
 from vcm.cloud import get_fs
 import fv3fit
-from ._plot_jacobian import plot_jacobian
+from ._plot_input_sensitivity import plot_jacobian, plot_rf_feature_importance
 from ._metrics import compute_metrics
 from ._mapper import PredictionMapper
 from ._helpers import (
     load_grid_info,
     sample_outside_train_range,
     is_3d,
+    get_variable_indices,
 )
 from ._select import meridional_transect, nearest_time
 
@@ -402,6 +403,7 @@ def main(args):
     batches = loaders.batches.batches_from_mapper(
         pred_mapper, variables, timesteps=timesteps, training=False, **batch_kwargs,
     )
+
     # compute diags
     ds_diagnostics, ds_diurnal, ds_scalar_metrics = _compute_diagnostics(
         batches, grid, predicted_vars=config.output_variables
@@ -419,7 +421,13 @@ def main(args):
     try:
         plot_jacobian(model, args.output_path)  # type: ignore
     except AttributeError:
-        pass
+        try:
+            input_feature_indices = get_variable_indices(
+                data=pred_mapper[timesteps[0]], variables=model._input_variables
+            )
+            plot_rf_feature_importance(input_feature_indices, model, args.output_path)
+        except AttributeError:
+            pass
 
     # compute transected and zonal diags
     snapshot_time = args.snapshot_time or sorted(timesteps)[0]
